Editor Page Refactoring
