+++ src/components/features/application/editor/SlideEditor.tsx
@@ -74,6 +74,7 @@ export default function SlideEditor({
         onUpdate(editor.getHTML());
       }
     },
+    immediatelyRender: false,
   });
 
   useEffect(() => {
