+++ src/components/features/application/editor/data/sampleSlides.ts
@@ -1,0 +1,212 @@
+// src/components/features/application/editor/data/sampleSlides.ts
+import { Slide } from "../EditorContainer";
+
+export const sampleSlides: Slide[] = [
+  {
+    id: 1,
+    title: "Can We REALLY Stop Climate Change?",
+    subtitle:
+      "The future of our planet is at stake. Let's explore the science, the impacts, and the potential solutions to climate change.",
+    author: "Talha Sabri",
+    editedTime: "about 1 month ago",
+    image: "/climate-earth.jpg",
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+    textColor: "#FFFFFF",
+    content: [],
+  },
+  {
+    id: 2,
+    title: "The Science: Understanding the Greenhouse Effect",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value:
+          "Greenhouse gases like carbon dioxide trap heat in our atmosphere",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value:
+          "Human activities, particularly burning fossil fuels, increase these gases",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "This warming effect is causing global temperature rise",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+  {
+    id: 3,
+    title: "Current Impacts: What We're Already Seeing",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value: "Rising sea levels threatening coastal communities",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value: "More frequent and intense extreme weather events",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "Disruption of ecosystems and biodiversity loss",
+      },
+      {
+        id: "content-4",
+        type: "text",
+        value: "Agricultural challenges and food security risks",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+  {
+    id: 4,
+    title: "Projections: Scenarios and Their Implications",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value: "Best case: Limiting warming to 1.5°C requires immediate action",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value:
+          "Business as usual: 3-4°C warming with catastrophic consequences",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "Tipping points could trigger irreversible changes",
+      },
+      {
+        id: "content-4",
+        type: "text",
+        value: "Economic costs increase dramatically with delay",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+  {
+    id: 5,
+    title: "Solutions: Addressing the Challenge",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value: "Transition to renewable energy sources",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value: "Improve energy efficiency across sectors",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "Develop carbon capture technologies",
+      },
+      {
+        id: "content-4",
+        type: "text",
+        value: "Protect and restore natural carbon sinks",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+  {
+    id: 6,
+    title: "Policy Approaches: Government Action",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value: "Carbon pricing mechanisms",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value: "Regulatory standards for emissions",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "Investment in clean technology research",
+      },
+      {
+        id: "content-4",
+        type: "text",
+        value: "International cooperation and climate agreements",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+  {
+    id: 7,
+    title: "Individual Impact: What Can We Do?",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value: "Reduce personal carbon footprint",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value: "Advocate for climate policies",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "Support sustainable businesses",
+      },
+      {
+        id: "content-4",
+        type: "text",
+        value: "Community engagement and education",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+  {
+    id: 8,
+    title: "Path Forward: Challenges and Opportunities",
+    content: [
+      {
+        id: "content-1",
+        type: "text",
+        value: "Balancing economic development with environmental protection",
+      },
+      {
+        id: "content-2",
+        type: "text",
+        value: "Ensuring just transition for vulnerable communities",
+      },
+      {
+        id: "content-3",
+        type: "text",
+        value: "Technological innovation as both challenge and opportunity",
+      },
+      {
+        id: "content-4",
+        type: "text",
+        value: "Building resilience while pursuing mitigation",
+      },
+    ],
+    backgroundColor: "#000000",
+    gradient: "linear-gradient(135deg, #111111 0%, #333333 100%)",
+  },
+];
